+++ Pipeliner/Model/Project.swift
@@ -1,0 +1,12 @@
+//
+//  Project.swift
+//  Pipeliner
+//
+//  Created by Michal Sverak on 26.08.2022.
+//
+
+import Foundation
+
+struct Project: Codable {
+    let name: String
+}
